Extracted closely related code as classes and divided classes into packages
