+++ src/main/java/Burrito/TaskList/Deadline.java
@@ -1,3 +1,5 @@
+package Burrito.TaskList;
+
 public class Deadline extends Task{
     protected String description;
     protected boolean isDone;
