+++ src/main/java/Burrito/TaskList/Event.java
@@ -1,3 +1,5 @@
+package Burrito.TaskList;
+
 public class Event extends Task{
     protected String description;
     protected boolean isDone;
