+++ src/main/java/Burrito/TaskList/Task.java
@@ -1,3 +1,5 @@
+package Burrito.TaskList;
+
 public class Task {
     protected String description;
     protected boolean isDone;
