+++ src/main/java/Burrito/TaskList/Todo.java
@@ -1,3 +1,5 @@
+package Burrito.TaskList;
+
 public class Todo extends Task{
     protected String description;
     protected boolean isDone;
